requirements.txt
alembic==1.5.8
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.1.0
importlib-metadata==4.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.4.9
typing-extensions==3.7.4.3
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.1
email_validator
PyJWT==1.4.2
Flask-JWT==0.3.2